gRPC/server-2 (Algorithms)/PFP/requirements.txt
numpy
matplotlib
scipy
protobuf
grpcio
requests
certifi